2019.06.11
내 1:1 문의내역

Board_1to1Service
Board_1to1ServiceImpl

Board_1to1Dao
Board_1to1DaoImpl 에 작성했던 코드

MemberService
MemberServiceImpl

MemberDao
MemberDaoImpl 로 옮김
