+++ Yagujango/src/dao/impl/MemberDaoImpl.java
@@ -5,6 +5,8 @@ import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 
 import dao.face.MemberDao;
 import dbutil.DBConn;
@@ -276,5 +278,55 @@ public class MemberDaoImpl implements MemberDao{
 		}
 		
 		return cnt;
+	}
+	
+	@Override
+	public List<Member> OneToOneSelectAll() {
+		
+		List<Member> OneToOneList = new ArrayList<Member>();
+		
+		String sql = "";
+		sql += "SELECT";
+		sql += " B.boardno,";
+		sql += " M.userno, M.userid, M.usernick,M.email,M.myteam,";
+		sql += " B.title, B.content, B.writer_comment, B.writtendate";		
+		sql += " FROM member M, board_1to1 B";
+		sql += " WHERE M.userid = B.writer_userid";
+		sql += " ORDER BY B.boardno";
+		
+		try {
+			ps = conn.prepareStatement(sql);
+			rs = ps.executeQuery();
+			
+			while( rs.next() ) {
+				Member member = new Member();
+				
+				member.setBoardno(rs.getInt("boardno"));
+				member.setUserno(rs.getInt("userno"));
+				member.setUserid(rs.getString("userid"));
+				member.setUsernick(rs.getString("usernick"));
+				member.setEmail(rs.getString("email"));
+				member.setMyteam(rs.getString("myteam"));
+				member.setTitle(rs.getString("title"));
+				member.setContent(rs.getString("content"));
+				member.setWriter_comment(rs.getString("writer_comment"));
+				member.setWrittendate(rs.getDate("writtendate"));
+				
+				OneToOneList.add(member);
+			}
+
+		} catch (SQLException e) {
+			e.printStackTrace();
+		} finally {
+			try {
+				if(rs!=null)	rs.close();
+				if(ps!=null)	ps.close();
+			} catch (SQLException e) {
+				e.printStackTrace();
+			}
+		}
+				
+		
+		return OneToOneList;
 	}
 }
