+++ Yagujango/src/service/impl/MemberServiceImpl.java
@@ -4,6 +4,7 @@ import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -120,6 +121,11 @@ public class MemberServiceImpl implements MemberService{
 			return false;
 		else
 			return true;
+	}
+	
+	@Override
+	public List<Member> getOneToOneList() {
+		return memberDao.OneToOneSelectAll(); 
 	}
 
 }
